requirements.txt
huggingface-hub==0.24.6
hf-transfer==0.1.8
python-dotenv==1.0.1